requirements.txt
Authlib==1.6.9
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
python-dotenv==1.2.2
Werkzeug==3.1.6
